# 06_Agent智能体合流总管/openapi/project4-medical-rag.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "医疗智能问诊基础RAG API - Dify 工具集",
    "description": "项目四：基于 Milvus 向量检索的医疗知识库问答接口",
    "version": "1.0.0",
    "contact": {
      "name": "系统管理员",
      "email": "admin@enterprise.com"
    }
  },
  "servers": [
    {
      "url": "http://host.docker.internal:9004",
      "description": "Docker 宿主机 - 项目四医疗RAG后端"
    },
    {
      "url": "http://localhost:9004",
      "description": "本地直接访问"
    }
  ],
  "paths": {
    "/api/chat": {
      "post": {
        "operationId": "medical_chat_query",
        "summary": "医疗知识库问答",
        "description": "查询医疗知识库（疾病诊疗、药物说明、科室介绍等），基于 Milvus 向量检索生成回答。当用户咨询医疗健康类问题时使用此工具。例如：'感冒了吃什么药'、'高血压的治疗方案'、'心内科看什么病'。注意：此工具仅用于医疗知识查询，不提供诊断建议。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["question"],
                "properties": {
                  "question": {
                    "type": "string",
                    "description": "用户的医疗健康问题",
                    "examples": ["感冒了吃什么药？", "高血压患者饮食注意事项"]
                  },
                  "top_k": {
                    "type": "integer",
                    "description": "返回相关文档片段数量，默认3",
                    "default": 3,
                    "minimum": 1,
                    "maximum": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "answer": {
                      "type": "string",
                      "description": "基于医疗文档生成的回答"
                    },
                    "sources": {
                      "type": "array",
                      "description": "引用的文档来源",
                      "items": {
                        "type": "object",
                        "properties": {
                          "doc_name": { "type": "string", "description": "文档名称" },
                          "chunk_text": { "type": "string", "description": "匹配的文本片段" },
                          "page_num": { "type": "integer", "description": "页码" },
                          "score": { "type": "number", "description": "相似度评分" }
                        }
                      }
                    },
                    "confidence": {
                      "type": "number",
                      "format": "float",
                      "description": "置信度 (0-1)"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "查询失败",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chat/stream": {
      "post": {
        "operationId": "medical_chat_stream",
        "summary": "医疗知识库流式问答",
        "description": "SSE 流式医疗知识库问答，逐字返回结果。用于需要实时展示回答的场景。",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["question"],
                "properties": {
                  "question": {
                    "type": "string",
                    "description": "用户的医疗健康问题"
                  },
                  "top_k": {
                    "type": "integer",
                    "default": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE 流式响应",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "description": "事件格式: event: sources/chunk/done, data: JSON"
                }
              }
            }
          }
        }
      }
    }
  }
}